Feature/secure storage 2 (#5991)
Adding REST API to access the secure store. 
Added Guice modules for secure store. This will need to be updated to return KMS based store once that is implemented.
Fixes a bug where the store initialization fails if the parent directory does not exist.

This is part of https://issues.cask.co/browse/CDAP-3906.