CDAP-6507 fix hydrator phase splitting bug
Fixed bugs with phase splitting introduced when we added support
for multiple inputs. After that change, the previous algorithm
of splitting the dag on connector nodes could result in subdags
that contained multiple reduce nodes.

Changed the algorithm to ensure that each subdag contains at most
a single reduce node.