HYDRA-388 add metrics for records emitted by streaming sources
Many stages in data streams were not emitted records in and records
out of the stage. Fixed by creating a CountingFunction that is used
before and after these stages that just counts records for that
stage. Also took this opportunity to fix HYDRA-364 using the function
instead of calling count() on rdds.