Updates the licenses and license dependency files. (#6496)
* Updates the licenses and license dependency files.
Adds error handling to the python scripts used for this. Handles the
case of an incorrect version of Python (requires 2.7 as a minimum for
building the PDFs) and missing apps such as npm and bower. Checks
if the copyright date included in the PDF generator is correct. (The
date can't be the current one as sometimes these are generated before
the change of the year.)

* updated license dependencies