HYDRATOR-180 document data streams
Updating docs with information about the data streams app.
Replaces some of the documentation on etl-realtime with information
about data-streams. Also adding documentation on the new
StreamingSource and Windower plugin types. Also added a little
blurb at the start of each plugin type section describing what
it is used for.