Update the CDAP dependencies for release 3.5.2.
Added a new option (-t) to the license tool that lists all libraries
that require special inclusion in the documentation because they are not
Apache License (v2), MIT, or Public Domain.
Added additional entries of documented licenses for linked or included
libraries.
Moved these to a new page.
Added links to licenses to the CDAP UI page.
Created new rst and pdf files.