CDAP-7181 Update Windows script (#6774)
* CDAP-7181 Update Windows script

* Changes to handle subroutines et al.

* Fixes script so that it works with Java paths and CDAP home paths
with spaces in them. Fixes the Java version test to work with either
Oracle JDK or OpenJDK. Adds the tx-debugger as an option. Adds a
progress bar (series of periods) while CDAP is starting. Adds notes
explaining how these issues were resolved in the script as appropriate.

* Update cdap.bat

Line up command usage.

* Restore cdap-cli.bat and convert to wrapper

* Update removal version, per #6902

* Support deprecated actions and remove CLI fall-through

* Fixes cdap-cli.bat to handle spaces in CDAP_HOME and corrects a path with incorrect separators.
Fixes cdap.bat with an instance of JAVA_HOME not guarded with double-quotes.