CDAP-2013 moving setting of program state into the program runners
This makes each program responsible for setting its state instead
of the ProgramLifecycleService in app-fabric. This is so that state
recording is located in a single place, and has the added benefit
that state will be properly recorded in the event that the cdap
master fails over while programs are running.

As part of this, also refactoring the program runners to consolidate
much of the common code.