CDAP-7683 add a tool to upload coprocessor jars
Added a tool that will build and upload coprocessor jars to HDFS.
The tool will run before the master starts, and can also be
run on its own. It can be used on a slave cluster to ensure that
required coprocessor jars are on HDFS even though CDAP is not
running on the cluster.

Refactored the logic around building coprocessors so that
HBaseTableAdmin and HBaseQueueAdmin no longer build their required
coprocessor on demand, but just assume the coprocessor exists at
a pre-defined location. Logic for building the actual jar was
just moved from HBaseTableUtil to a new CoprocessorManager class.