CDAP-7724 fix a race condition in stopping spark programs
Changed the program controller used for Spark and MapReduce to
wait for the underlying guava service to execute its listeners
before transitioning the controller to the killed state. Otherwise,
the service listeners may not complete before the twill runnable
finishes and shuts down. This manifests itself as a failure to
record the program state and perform other cleanup.