CDAP-8036 add a fat coprocessor jar to hbase compat packages
Added a fat jar containing all the CDAP coprocessors to every
hbase compat package in a new 'coprocessor' directory. It is
meant to be used in CDAP setups where the replication status
tool is required, or setups where 'master.manage.hbase.coprocessors'
is set to false in order to enable rolling hbase upgrades.