CDAP-8288 improvements to prebuilt coprocessors
Some refactoring and simplification to the pre-built coprocessors.
Instead of a coprocessor per table type, just using a single
coprocessor jar with all the coprocessor classes in them. This
makes it simpler and mirrors what would happen in a cluster
setup where the coprocessor jar is deployed to all regionservers.
Also some refactoring to move duplicate logic around the
CoprocessorDescriptor to a single place.

Another change was to remove the cdap build time from the
coprocessor path for non-snapshot versions of cdap. There is
also a setting that can be set to false to omit it entirely. This
is purely to make testing easier on clusters that have different
build times for the same cdap snapshot.

Also adding an option to the coprocessor build tool to overwrite
any jar that already exists.