HYDRATOR-1323 introduce error transform
An ErrorTransform can be placed after another stage to consume
the errors emitted by that stage. MapReduce changes boil down to
differentiating between stages that takes output and stages that
take errors. When emit() is called, it will only be sent to stages
that take output. When emitError() is called, it will only be sent
to stages that take errors.

Spark refactoring amounts to making stage return a Tuple2 of
boolean and object. The boolean indicates whether the second
object is an error record or an output record. Each stage that
can emit both is then filtered to make two RDDs, one that only
has outputs and one that only has errors.