CDAP-8032 retry when connect fails after service discovery
Refactoring several clients that perform service discovery
then http requests so that they use a single common class. This
client will throw ServiceUnavailableException if the service
cannot be discovered, if discovery succeeded but a ConnectException
was thrown while making the request, or if the response itself
was a 503 service unavailable.