CDAP-7302 retry getting table to prevent master from dying
The DatasetServiceStore is used to store and get the number of
instances for system services. This Service used to throw an
exception on startup if hbase was down. This situation can happen
if a master becomes leader while hbase is down. Changing it to
retry the operation so that the master does not die.