CDAP-8672 don't log errors in router for 503
Fixes a bug where 503 responses due to an undiscoverable service
would result in an error log and a stack trace. A 503 is normal
behavior for the router and does not indicate an error in the
router. As such, logging these as debug instead of error. Only
internal router errors should be logged as error.