CDAP-8744 fix coprocessor upgrade
Fix the conversion from HTableDescriptor into a TableDescriptor
so that it does not copy properties that represent coprocessors,
since the TableDescriptor has its own coprocessor fields.
This fixes a bug in the upgrade tool where tables with more than
one coprocessor would end up with a copy of each coprocessor.