CDAP-9287 add a method to get arguments at runtime for all plugins
Adding an Arguments class that SettableArguments extends,
which gives read access to all pipeline arguments for the run.
This will be accessible at runtime by all plugins.

Also fixed a bug where arguments set by an Action plugin would
not be visible to an Action that is run after it.

Also performing cleanup with the context classes. Combined some
context classes if possible, and removed a lot of redundant
information that they were passing along.