CDAP-9473 use api-spark-base as source
The spark and spark2 api modules will use the api-spark-base
module as source code in the dist profile, to ensure that the
base classes get compiled using the correct version of scala.

Also added a cdap-dev maven profile that adds cdap-api-spark-base
as a dependency of the api modules. This profile is purely
for use in IDEs, so that they don't get confused by two modules
using the same files as source.

CDAP-9473 skip deploy plugin for cdap-api-spark-base