CDAP-9473 move SparkMain and JavaSparkMain to base
Moving SparkMain and JavaSparkMain to cdap-api-spark-base, since
we do not expect them to diverge. As part of this, also removing
things that were only there for the IDE.

Removed the cdap-dev maven profile and removed cdap-api-spark-base
as an actual module, since it is just shared source for the
cdap-api-spark modules.