CDAP-9474 add a spark2 core module
Splitting spark-core into one module for spark1 and one for spark2.
Most of the current cdap-spark-core has been moved to
cdap-spark-core-base, which is used as source code for both
cdap-spark-core and cdap-spark-core2_2.11. The only classes in each
Spark specific module are classes that use some Spark API that
has changed.

Also changing sdk packaging so that both spark extensions are
packaged. Each ProgramRuntimeProvider is annotated with a
different spark version so that only the correct one is used at
runtime. In the SDK, users can switch spark versions by changing
a config setting. A future change will add logic in distributed
mode to discover the spark compat version so that the right
spark extension is used at runtime.