CDAP-9476 fix spark streaming in local mode
Spark2 streaming with checkpointing expects the job jar to be in
the same place in between runs, which means we can never delete
the jar. So in local mode, create the job jar in the app temp dir,
but not in the job specific temp dir.

Also some minor fixes to get preview tests to work.