CDAP-11609 fixes for spark2 on cdh
Modified spark submit to include deploy-mode as 'cluster' and
master as 'yarn', instead of master as 'yarn-cluster'.
This is to prevent issues that arise when an environment has a
spark default of 'client' set for deploy mode, which
clashes with the cluster in 'yarn-cluster'.

Also setting a property in hConf to make Hive use delegation tokens
instead of kerberos for auth in cdh environments.

Also explicitly filtering out the spark-yarn-shuffle jar that cdh
packages in their yarn lib directory, since it causes wrong classes
to be loaded.