CDAP-12037 implement multioutput for mapreduce
The TransformExecutor classes were refactored a bit to support
output ports. PipeTransformDetail was renamed to PipeStage, which
represents some stage in the pipeline. Some cleanup was done so
that these stages can be created all at once instead of
incrementally through a recursive traversal of the pipeline.

Instead of handling KeyValue objects, the classes have been
changed to handle RecordInfo objects, which contain actual record
values as well as context about where the record came from.
This allows each PipeStage to use whatever information it needs,
and will be re-used later in the Spark implementation.