CDAP-12152 mapred implementation for alert publisher
In MapReduce, an AlertPublisher is implemented as a sink that
writes to a local PartitionedFileSet. At the end of the
workflow, the workflow client will scan that fileset and publish
the alerts. The implementation is very similar to a
ConnectorSink, where an actual plugin is not instantiated in the
mapper/reducers, but a special class that knows how to write an
Alert and knows how to setup the PartitionedFileSet.