CDAP-12191 expose pipeline info for each plugin context
Adds methods to get pipeline name, namespace, logical start time,
and output port schemas to StageContext so that every plugin has
access to them at runtime. Also doing some cleanup and refactoring
so that exposing more information in the future will be easier.

Consolidated common pipeline info and CDAP contexts into a
PipelineRuntime class that is passed around in all the plugin
context constructors. The next time we need to expose more fields,
or CDAP capabilities, we can just modify that class instead of
modifying every single constructor in the context hierarchy.
Also removed as many of the constructors as possible to try and
simplify things.

Also removed the getHadoopJob method that has been deprecated since
3.x and which usually just throws an UnsupportedOperationException.