CDAP-12167 pipeline explicit transactions
Refactoring to make the initialize and destory methods in
ETLMapReduce, ETLSpark, and SmartWorkflow use explicit transaction
control. This is an improvement so that we don't run into
transaction timeouts unless a single plugin takes longer than the
short tx timeout. It also required in case alert publishing takes
longer than the short tx timeout.

Refactored the relevant plugin contexts to use the DatasetContext
provided by a TxRunnable instead of the MapReduceContext or
SparkClientContext as the DatasetContext. In order to do this,
added a context provider class that will return a new plugin
context given a DatasetContext.

As part of the refactoring, fixing CDAP-10974 was almost free,
so added a method to get SettableArguments to StageSubmitterContext