CDAP-12122 fix control dag flatten bug
Fixed a bug in the flatten logic of the control dag that could
create a flattened dag without a join node, which breaks the logic
in SmartWorkflow. The bug happens when there is a branch and
some of the branches join to a single node, but other branches
end in a sink. In those situations, a join node was not being
inserted, causing issues.