CDAP-12125 fix mapred pipeline timing metrics
Fixing a bug in pipeline timing metrics for mapreduce pipelines.
Instead of timing the total time spent in transform methods, ignoring
time spent emitting a record. The reason is that in mapreduce, an
emit call will end up calling the transform methods of subsequent
stages. This means stages early in the pipeline end up counting
time spent in later stages.