CDAP-12111 allow plugins to have their own plugins
Changes the plugin model so that a plugin can itself export some
api classes that other plugins can implement. For example, this
will allow a transform for the data-pipeline app to define its own
interface that other plugins can implement.

At configure time, when a plugin declares that it uses another
plugin, the platform will first try to find a plugin that extends
the calling plugin before finding a plugin that extends the
application. Capping it to just two levels for now to avoid some
possible complex dependency graphs and limit the number of lookups
required. When a plugin of a plugin is instantiated, its parent
classloader will be the classloader of the parent plugin.

CDAP-12111 some code cleanup