CDAP-12438 add connector support for spark pipelines
Enhancing spark pipeline programs so that they can handle
connector plugins. They never used to have to worry about
connectors because they could do aggregations and joins in a
single spark program. But with the addition of conditions,
spark programs can have connectors as input and output as well.

As part of this, refactoring the ConnectorSource and ConnectorSink
so that there are two implementations. The first handles records
of different types and is used in mapreduce. The second is simpler
and handles writing just a single type of object and is used in
Spark, since Spark connectors are only used for conditions.