CDAP-12501 fix schema check for conditions with multiple input types
A partial fix to the check that the input schemas into a condition
node must all be the same. Right now it checks that all non-null
input schemas are the same, but the full fix should be to check
that all non-action input schemas are the same. However, doing
that would require some refactoring that should not be done at
this time.