CDAP-12553 use correct context classloader inside a tx runnable
MapReduce and Spark program runtime services take care to set
the context classloader to a special classloader that exposes
packages exported by plugins. This special classloader was
being overwritten when running a TxRunnable, where it was being
set to a CombineClassLoader of the program classloader and the
CDAP classloader. This would cause issues if plugins run inside
a TxRunnable were trying to load classes using the context
classloader, which various Hadoop formats will do.