CDAP-12946 fix the pipeline action planner
Fixing a bug that caused pipeline actions to form incorrect
workflows. In some scenarios, actions could cause pipelines to
fail to deploy. In other scenarios, actions could cause some
parts of the pipeline to be excluded from execution.

These issues usually arise when there are multiple sources and
an action is placed in front of the sources.

The problem was in the logic that splits up a pipeline based on
the control nodes (actions are a control node). The planner was
incorrectly creating a subdag for each source connected to an
action, instead of a single subdag for all sources connected to
the action.