CDAP-13120 use correct directory for file in PartitionedFileSet
If there was no base path set for a PartitionedFileSet, it would
default to a 'files' directory whose parent is a directory named
after the dataset. This means when the dataset is deleted, only
the 'files' directory is deleted, but not the parent directory.
This behavior is especially problematic when using a
PartitionedFileSet as a local fileset in a workflow, as each run
of the workflow would leave a directory per PartitionedFileSet.