CDAP-13096 record program run cluster state
Changing the state writer to listen for cluster state
events on TMS and record the state change. Also adding some fixes
for program state when a run goes from provisioning to deprovisioning
or deprovisioned due to some failure in provisioning.

Also remove startTime as an argument to the method for recording
provisioning state since the start time is encoded in the runid.