CDAP-13096 refactor program lifecycle to trigger off TMS
Refactoring so that program execution will be triggered off events
on TMS in preparation for additional cluster lifecycle states.
Refactoring the ProgramLifecycleService to remove usage of
ProgramController when possible, as starting a run will first need
to provision the runtime before the ProgramController can be
used.

Added a 'pending' program run status to indicate the actual
program has not been started. This corresponds to the provisioning
cluster state. Also adding a 'starting' program state to the
status rest endpoint, as a 'running' status when a cluster is
being provisioned is quite misleading, and breaks a bunch of
unit tests that assume a 'running' program status means everything,
like service instances, are actually up and running.
Along with this change, fixing unit tests to define isRunning
as anything that is not stopped.