CDAP-13096 move scheduler out of program lifecycle service
Moving Scheduler out of ProgramLifecycleService to avoid a
circular dependency chain with ProgramLifecycleService to
CoreSchedulerService to ConstraintCheckerService back to
ProgramLifecycleService.

Also move direct usage of Scheduler that was taking place in the
ProgramLifecycleHttpHandler into a new ProgramScheduleService.
The remaining logic in the handler is validation of user input,
transformation of user string input into internal classes, and
transformation of internal classes to what should be returned
by the REST endpoints.

Also doing various cleanup and bug fixes with scheduler operations.
For example, authorization was not being performed when getting
schedule details, but was when getting schedule status.

Also starting ProgramLifecycleService in TestBase in preparation
for cluster lifecycle changes, and to make sure that
ProgramLifecycleService can be instantiated.