deprecate and replace waitForStatus() calls in unit tests
This method is almost always subject to race conditions, where
a program is starting but not yet running, and other parts of the
unit test expect it to be running. The existing race conditions
will become much more likely once the additional provisioning
lifecycle stages are introduced, which causes programs to get to
the running state a little slower now.