replace waitForStatus(true) calls in unit tests
This method is almost always subjected to race conditions, where
a program is starting but not yet running, and other parts of the
unit test expect it to be running. The existing race conditions
are made much more likely due to the additional provisioning
lifecycle stages, which causes programs to get to the running
state a little slower now.