CDAP-13328 wait for scheduler to start in AppFabricTestBase
Also forcing tests that extend the class to pass in the expected
response code when deploying an application. There were a bunch
of test cases that try to deploy an app, then assume it was
successfully deployed. If it actually failed, some other assertion
in the test would fail, making it difficult to diagnose what
actually happened. Now, if the response code does not match, the
assertion failure message will include the response message from
the deploy call, making it very clear why the test failed.