CDAP-13463 fix pipeline planner when conditions are on branches
SmartWorkflow had a bug where it assumed that all programs
on a branch would not fork or be conditions. It is still true
that they won't fork, but it has never been true that they cannot
be conditions. Fixing to handle that scenario, adding unit tests,
and performing general cleanup to fix warnings.