CDAP-13493 remove operational stats
Operational stats were found to be not useful for most people,
while increasing the likelihood of generating misleading warning
logs for certain cluster configurations.

For example, secure clusters with HA namenode would cause the
HDFS stats to log an authorization error that confused most users.

Operational stats are also confusing when used in conjunction with
the cloud integrations planned for 5.0.0.