CDAP-13489 re-fix dynamic partitioning test
The previous fix was not correct, as it was too aggressive in
removing logic. The call to start a program does not guarantee
any state change by the time the call returns. We therefore cannot
just call waitForStopped() right after calling start().

We really need to change the way a program run is handled in
unit tests.