CDAP-13501 fix flaky run history tests
Fix a race condition where a call to list run records is made first,
then calls to fetch individual run records are made and the results
compared. This is subject to race conditions where the cluster state
has changed. For example, when listing, a run record might be in the
DEPROVISIONING state. When it is individually fetched, it could be
in the DEPROVISIONED state. Instead of comparing entire run records,
just comparing run record program state, since that is really all that
is important for the test.

Also performing cleanup to remove duplicate code and to remove a bunch
of warnings.