CDAP-13489 another dynamic partioning test fix
The previous fix can fail due to a race in program status
(CDAP-13296) where run records are not fetched in a single
transaction. If a run transitions state during a status call,
the status can incorrectly be returned as STOPPED.
Instead, just wait for the number of failed runs to increase,
which should always indicate that the run has completed.