CDAP-13358 rename 'default' profile to 'native'
This rename is to avoid confusion, since the word default has
meaning about how a profile is used, but not what the profile is.
For example, the default profile for a namespace is actually
whatever profile is set for that namespace, which might be a user
created profile. Changing it to 'native' so that it describes
what the profile does.