CDAP-13604 cleanup around the MetadataDataset
Added javadocs to the MetadataDataset.

Moved search arguments into a SearchRequest class so that we can
pass around a single object instead of ten arguments everywhere.
This also allows us to perform basic validation on every request
in a single place, early on in a request, instead of waiting until
we get down to the dataset.

Also added some unit tests for DefaultValueIndexer