CDAP-13604 add cross namespace metadata search
Cross namespace search is implemented by adding additional index
columns that don't contain prefix the index value with the entity
namespace. When index value 'foo' is written on an entity in
namespace 'ns1', instead of just writing 'ns1:foo' to column 'i',
we will also write 'foo' to column 'xi'.

This adds new indexes, and will rely on upgrade to add those new
indexes.

Performing some refactoring to the MetadataDataset by introducing
a SearchTerm class that holds a namespace, term, and whether
it is a prefix search. Using the class instead of performing
string manipulation everywhere.