CDAP-13673 use fileset instead of partitioned fileset for connectors
Changing to use fileset to remove the usage of a Table. This will
allow pipelines that use connectors to work in cloud environments

Also fixing a bug in FileSetAdmin and Locations, where it would
recursively check for parents of the fileset location. In cloud
environments, this would go all the way up to the root location
and result in an exception because there would be no path in the
location URI. Instead, changing FileSetAdmin to stop the recursive
parent lookup when it gets to one level above the root directory.