CDAP-13614 fix bug with multiple deprovisioning notifications
The root cause was that we were double writing run state if the
multiple run state messages were sent. When a program is killed,
the CDAP master sends a killed message and so does the program.
Both of these would trigger a deprovisioning message, and the
second message would cause errors because a deprovisioning task
had already run.

Fixing by avoiding the run state double write.