CDAP-14104 add ability to determine which ip to use for dataproc
Changed the Dataproc provisioner so that if it is running on GCP
in the same network as the dataproc cluster, it will by default
use the internal ip address when communicating with the cluster.
This makes it less likely that firewall rules will be an issue.
If users still want to use the external ip address, that is left
as an option.

Performed some refactoring so that CDAP no longer looks for the
'ip.external' node property, since that is no longer accurate.
Instead, added ip address as a required field in the Node class,
since it was functionally required anyway. This allows us to make
the contract explicit, instead of having required properties in
some Map.

Also made network and zone optional properties. When the provisioner
is running on GCP, the network and zone can be picked up from the
metadata server running on the VM.