CDAP-14022 add structure for bootstrap framework
A new 'bootstrap.file' config setting was added, which is expected
to point to a file that contains a list of bootstrap steps to
perform, such as loading system artifacts and creating the default
namespace. Each type of step expects different arguments in a
different format.

Added a bootstrap framework. Each type of step has its own
BootstrapStepExecutor with its own arguments class. A BootstrapService
is responsible for reading a bootstrap file and performing each
step on CDAP startup.

Modified the build to contain a bootstrap file.