CDAP-14161 fix flaky cancel deprovision test
In the provisioning tests, we need to call the deprovision method
that lets the caller specify what sort of task cleanup to perform.
This is to prevent the default behavior, which is to delete task
state after a deprovision task is complete, which would not allow
the test to verify state.