CDAP-14498 improve app deployment failure messages
Removed the java exception classname from the error message by
properly unwrapping an ExecutionException in the depolyment code.

Also shortening an error message in the pipeline validation logic
to only include stage name instead of plugin type and name, as
the plugin type and name are not exposed to end users.