CDAP-14795 add namespaced create classloader method
Added a namespaced version of the create classloader method in
ArtifactManager to mirror the namespaced version of the list
artifacts method.

This is required when an app in one namespace lists artifacts
in another namespace and wants to create a classloader. If the
namespace is not provided in the method, the namespace of the app
will be used, which will not be the true namespace of the artifact,
resulting in that artifact not being found.