CDAP-14813 fix flaky explore service tests
The queries used to create and drop databases were being treated
as blocking calls in the unit tests even though they are not.
Changed them to wait until the query is done before moving on.

This fixes one particular test that would create namespace 'foo',
check for TableNotFoundExceptions on two tables in 'foo', then delete
namespace 'foo'. The exceptions can be thrown if the namespace
does not exist, and the deletion of the namespace can fail if
it is run before the creation of the namespace completes. This
leaves namespace 'foo' hanging around for the lifetime of the
test, which causes a subsequent namespace test case to fail.