CDAP-14839 replace usage of NotFound and AlreadyExists exceptions
Replaced NotFoundException and AlreadyExistsException with classes
specific to the storage SPI. This was done to prepare to move
the SPI definitions into their own module with minimal dependencies.
NotFoundException and AlreadyExistsException are in cdap-common,
which cannot be used.

Also corrected some exception handling in the ArtifactStore if
getting a StructuredTable throws a TableNotFoundException.
It was not correct to throw a NotFoundException, as that type of
error should not result in a 404, but a 500 response code.

Also removed some guava usage, as the SPI should not depend on
guava either.