CDAP-14894 migrate config store to use new SPI
The ConfigDataset used to be used for the DefaultConfigStore
and the PreferencesService. However, the PreferencesService
has a bug that makes it impossible to move it to a sql backed
implementation without changing the way preferences are stored.
Preferences are also tied together with profile and other stores.

To workaround this, renamed the ConfigDataset to
InternalPreferencesDataset and made it only used by the
PreferencesService. The DefaultConfigStore was moved to use the new SPI,
so config and preferences no longer share the same underlying storage.