CDAP-14905 add program runtime service for kubernetes
Added a kubernetes ProgramRuntimeService that only works with
service programs for the time being.

To run a program, it first creates a config-map for the run that
contains the application spec and the program options. It then
creates a deployment that has the config mounted on a volume.
The service program knows how to read those configs in order
to run the program.

Some classes were refactored in order to support these changes.
A SerDe interface was added in order to allow the Kubernetes
runner to serialize the app spec and ProgramOptions instances
without access to gson and the relevant codes located in app-fabric.

The AbstractProgramController was also moved into the master-spi,
as it contains the base logic for any program controller. In order
to move it, guava usage was removed. ListenableFuture was replaced
by CompletableFuture and equivalent utility methods from guava
were added to the class.