CDAP-15599 log dataproc creation warnings
Log warnings that are encountered when create dataproc clusters.
Also change the default disk sizes so that they don't generate
a warning about disk performance.

Also add some logic when a delete requests to check if the cluster
state is already in the deleting state. This is done to avoid
logging a red herring in a specific edge case.