CDAP-15747 fix a race when cancelling a KeyedExecutor task
Fixed a race condition in KeyedExecutor where the task could
get cancelled but return that it was not cancelled.
This fixes a race condition in the provisioner where it can
cancel a task, think that the task completed before it was
cancelled, then skip the logic to modify the run state.