CDAP-15943 fix remote runs when master is using internal ssl
For remote runs, unset the ssl cert path so that it doesn't
try and use a cert/private key file that doesn't exist.

Also update the RemoteExecutionDiscoveryService to serialize
an entire Discoverable instead of just the address in order to
preserve the payload, and to correctly handle http/https based
on configuration.