Update commons-collections version to 3.2.2
The commons-collections v3.2.1 has a vulnerability which was fixed in v3.2.2. Refer: https://commons.apache.org/proper/commons-collections/release_3_2_2.html
CDAP is using hadoop-common version (2.3.0) and hbase-common (1.0.1.1) in pom which is pulling the commons-collections version 3.2.1. I've excluded the transitive dependency and added the commons-collections v3.2.2 as a direct dependency in the pom.