CDAP-17196 fix mapreduce on distributed mode
mapreduce was failing due to a missing asm-tree dependency.
Somehow this dependency was not getting traced since the upgrade
to asm 7, even though it is a twill dependency.

Added MapReduceContainerLauncher as a class to trace for
dependencies when building the job jar since it is used to
launch the actual mapreduce job.

Also modified the launcher to print unexpected errors to stderr
instead of using the logger, to ensure that errors are not lost.