[CDAP-17246] Fixes pipeline studio to support pipelines exported in older version of CDAP (<6.2) to not change stage names.
The fix involves using a unique ID that will be used in DOM node. The reason we changed the name in
the first place was to have no spaces (or other spl characters) that are not supported by DOM id attribute.
Since the same stage name could be used in other plugin properties we no longer change the name during pipeline
imports.

We however have to update the stage name when copy/pasting nodes in the pipeline studio. We need to do this as
the stage name uniquely identifies the plugin in the pipeline DAG. This change makes sure stage names modified
while copy/pasting nodes to pipeline canvas is correspondingly updated in other plugins where the plugin
name is used.