CDAP-17389 improvements to dataproc provisioner
Fixed the provisioner to use the same worker disk settings for
secondary workers, and to use non-preemptible secondary workers.
This makes autoscaling reliable instead of causing pipelines
to fail. Also added a property for the number of secondary
worker nodes since some autoscale policies require a minimum
number of secondary nodes. This required bumping the dataproc
library version and switching from v1beta2 api to v1.

Added a disk type property for master and worker nodes to support
creating clusters with SSDs.

Fixed the widget for Component Gateway to correctly send
'true' and 'false' instead of 'on' and 'off' to the backend.

Changed all boolean properties to toggles for consistency.

Removed 'Beta' and 'Experimental' from descriptions/lables.

Moved a bunch of optional properties to the 'Advanced' section.